src/test/kotlin/swali/rules/zalando/HyphenateHttpHeadersRuleTest.kt
package swali.rules.zalando

import io.swagger.models.Swagger
import org.assertj.core.api.Assertions.assertThat
import org.junit.Test
import swali.*

class HyphenateHttpHeadersRuleTest {
    val rule = HyphenateHttpHeadersRule(config.headersWhiteList)

    @Test
    fun simplePositiveCase() {
        val swagger = swaggerWithHeaderParams("Right-Name")
        assertThat(rule.validate(swagger)).isNull()
    }

    @Test
    fun simplePositiveCamelCase() {
        //TODO MK: fix it
        // CamelCaseName IS a valid 'hyphenated' header, it just has a single term
        val swagger = swaggerWithHeaderParams("CamelCaseName")
        assertThat(rule.validate(swagger)).isNull()
    }

    @Test
    fun mustAcceptValuesFromWhitelist() {
        val swagger = swaggerWithHeaderParams("ETag", "X-Trace-ID")
        assertThat(rule.validate(swagger)).isNull()
    }

    @Test
    fun emptySwaggerShouldPass() {
        val swagger = Swagger()
        assertThat(rule.validate(swagger)).isNull()
    }

    @Test
    fun positiveCaseSpp() {
        val swagger = getFixture("api_spp.json")
        assertThat(rule.validate(swagger)).isNull()
    }

    @Test
    fun positiveCaseTinbox() {
        val swagger = getFixture("api_tinbox.yaml")
        assertThat(rule.validate(swagger)).isNull()
    }

    @Test
    fun issue572RateLimitHeadersAreAccepted() {
        val swagger = swaggerWithHeaderParams("X-RateLimit-Limit", "X-RateLimit-Remaining", "X-RateLimit-Reset")
        assertThat(rule.validate(swagger)).isNull()
    }
}
